Added router and routes
